backend/requirements.txt
# FastAPI server
fastapi==0.115.0
uvicorn==0.30.6
asyncpg==0.30.0
python-dotenv==1.0.1

# Pipeline scripts (crypto_terminal/)
psycopg2-binary==2.9.9
requests==2.32.3
feedparser==6.0.11
numpy==1.26.4
pandas==2.2.2
torch==2.4.0
